Remove url encoding
I've been getting a number of messages that the previous method was resulting in messages like "SubscribeMe!+/r/nosleep+/u/username" when the link is clicked on an android browser. Url encoding is useful when you're expecting non-ascii characters, but everything there is safe and usernames are always ascii.

If you want to keep the encoding, there are a few alternatives, but I think this is the simplest approach for this.